02_LLM/LLM_local_api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
requests>=2.28.0
prometheus-client>=0.16.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=0.21.0
Jinja2>=3.0.0
pdf2image>=1.16.0
pillow>=9.0.0
PyYAML>=6.0
faiss-cpu>=1.7.0
langchain-community>=0.0.0
langchain-text-splitters>=0.0.0
openai>=1.32.0
pypdf>=4.0.0 